Function on chef solo correctly
Explicitly access data bags rather than relying on search which is not available on chef solo without chef-solo-search. We'll eventually provide LWRPs for users and dialplan contexts, or allow downstream cookbooks to drop their own config for these, so complex search can be used in downstream cookbooks to trigger these.